Closes #10: Implement option variables. Closes #28: Implement debug command-line parameter
This commit did check and make sure that option variables are implemented correctly.

  - Variable definition is made public in pmake_variables
  - Renamed _ON / _OFF constants to _ON_ and _OFF_
  - Some minor fixed on the OutputLn, StdErrLn, StdOutLn functions in the code
  - The force build option was not handled from the commandline by make
  - pmake-gui: replaced treeview by valuelist editor
  - pmake-gui: reading and writing to PMakeCache.txt is implemented

+++ pmake/make_main.pas
@@ -21,11 +21,12 @@ type
   TRunMode = (rmBuild, rmInstall, rmClean);
 
 const
-  CmdOptions: array[1..7] of TCmdOption = (
+  CmdOptions: array[1..8] of TCmdOption = (
     (name: 'build'; descr: 'Build all targets in the project.'),
     (name: 'clean'; descr: 'Clean all units and folders in the project'),
     (name: 'install'; descr: 'Install all targets in the project.'),
     (name: '--compiler'; descr: 'Use indicated binary as compiler'),
+    (name: '--debug'; descr: 'Do not delete the make2 source file.'),
     (name: '--force'; descr: 'Force building make2.exe'),
     (name: '--help'; descr: 'This message.'),
     (name: '--verbose'; descr: 'Be more verbose.')
@@ -34,6 +35,7 @@ const
 var
   sline: TStringList;
   force_build: boolean = False;
+  debug: boolean = False;
   RunMode: TRunMode = rmBuild;
 
 procedure output_line(var sline: TStringList);
@@ -205,7 +207,9 @@ begin
   if verbose then
     OutputLn('-- Deleting temporary files');
 
-  DeleteFile(src_name);
+  if not debug then
+    DeleteFile(src_name);
+
   DeleteFile(ChangeFileExt(src_name, '.o'));
 
   if exit_code <> 0 then
@@ -269,6 +273,8 @@ begin
                 usage;
               end;
             end;
+            '--force': force_build := True;
+            '--debug': debug := True;
             '--help': usage;
             '--verbose': verbose := True;
           end;
